Moved internal source.c functions into new source_state module
