+++ src/sacn/common.c
@@ -21,6 +21,7 @@
 
 #include "sacn/private/mem.h"
 #include "sacn/private/source_loss.h"
+#include "sacn/private/source_state.h"
 #include "sacn/private/sockets.h"
 #include "sacn/private/source.h"
 #include "sacn/private/receiver.h"
@@ -73,6 +74,7 @@ etcpal_error_t sacn_init(const EtcPalLogParams* log_params)
     bool mem_initted = false;
     bool sockets_initted = false;
     bool source_loss_initted = false;
+    bool source_state_initted = false;
     bool receiver_initted = false;
     bool source_initted = false;
     bool merger_initted = false;
@@ -101,6 +103,8 @@ etcpal_error_t sacn_init(const EtcPalLogParams* log_params)
     if (res == kEtcPalErrOk)
       source_loss_initted = ((res = sacn_source_loss_init()) == kEtcPalErrOk);
     if (res == kEtcPalErrOk)
+      source_state_initted = ((res = sacn_source_state_init()) == kEtcPalErrOk);
+    if (res == kEtcPalErrOk)
       receiver_initted = ((res = sacn_receiver_init()) == kEtcPalErrOk);
     if (res == kEtcPalErrOk)
       source_initted = ((res = sacn_source_init()) == kEtcPalErrOk);
@@ -128,6 +132,8 @@ etcpal_error_t sacn_init(const EtcPalLogParams* log_params)
         sacn_source_deinit();
       if (receiver_initted)
         sacn_receiver_deinit();
+      if (source_state_initted)
+        sacn_source_state_deinit();
       if (source_loss_initted)
         sacn_source_loss_deinit();
       if (sockets_initted)
@@ -162,6 +168,7 @@ void sacn_deinit(void)
     sacn_dmx_merger_deinit();
     sacn_source_deinit();
     sacn_receiver_deinit();
+    sacn_source_state_deinit();
     sacn_source_loss_deinit();
     sacn_sockets_deinit();
     sacn_mem_deinit();
