+++ src/sacn_mock/sources.cmake
@@ -3,5 +3,6 @@
 set(SACN_MOCK_SOURCES
   ${SACN_SRC}/sacn_mock/common.c
   ${SACN_SRC}/sacn_mock/source_loss.c
+  ${SACN_SRC}/sacn_mock/source_state.c
   ${SACN_SRC}/sacn_mock/sockets.c
 )
