+++ tests/unit/api/cpp/merger/CMakeLists.txt
@@ -10,10 +10,10 @@ set(TEST_MERGER_SOURCES
   ${SACN_SRC}/sacn/merge_receiver.c
   ${SACN_SRC}/sacn/source.c
   ${SACN_SRC}/sacn/util.c
-  ${SACN_SRC}/sacn_mock/common.c
-  ${SACN_SRC}/sacn_mock/source_loss.c
+  
   ${SACN_SRC}/sacn_mock/dmx_merger.c
-  ${SACN_SRC}/sacn_mock/sockets.c
+  
+  ${SACN_MOCK_SOURCES}
 )
 
 sacn_add_dynamic_test(test_cpp_merger ${TEST_MERGER_SOURCES})
